+++ Services/RequestServer/RequestClient.ipc
@@ -1,5 +1,5 @@
 #include <LibHTTP/HeaderMap.h>
-#include <LibRequests/NetworkErrorEnum.h>
+#include <LibRequests/NetworkError.h>
 #include <LibRequests/RequestTimingInfo.h>
 #include <LibURL/URL.h>
